fixed the error of calling "materials"

+++ app/src/main/java/ru/android/hyrulecompendiummvp/app/models/data/HyruleData.kt
@@ -24,5 +24,5 @@ class HyruleData(
     @SerializedName("cooking_effect")
     val cookingEffect: String?,
     @SerializedName("hearts_recovered")
-    val heartsRecovered: Int?
+    val heartsRecovered: Double?
 )
